Add file check to config installation (.dotfiles/etc/home/init.sh)

+++ etc/home/init.sh
@@ -7,7 +7,15 @@
 DESTINATION_PATH=$HOME
 
 configure() {
+	# Check if a files, directory, link, etc already exists
+	if [ -e "$DESTINATION_PATH/$2" ]
+	then
+		echo "Cannot write to $DESTINATION_PATH/$2"
+		exit 1
+	fi
+
 	ln -s "$PWD/$1" "$DESTINATION_PATH/$2"
+	echo "Successfully linked $DESTINATION_PATH/$2 to $PWD/$1"
 }
 
 case "$1" in
@@ -22,5 +30,6 @@ case "$1" in
 		;;
 	*)
 		echo "Unknown option"
+		exit 1
 		;;
 esac
